pyproject.toml
[project]
name = "graphql-authz-proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "graphql-core>=3.2.6",
    "gunicorn>=23.0.0",
    "jsonpath-ng>=1.7.0",
    "pydantic>=2.11.9",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "typer>=0.17.4",
]
[dependency-groups]
dev = [
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.2",
    "ruff>=0.13.1",
    "testcontainers>=4.13.0",
]

[project.scripts]
gqlproxy = "graphql_authz_proxy.cli:typer_app"

[tool.ruff]
line-length = 120
exclude = [".venv", "__pycache__", "build", "dist", "tests", "identity_providers"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "C90", "D", "N", "S", "ANN", "COM", "T20", "Q", "RSE", "SIM", "ARG", "PL", "TRY", "ERA", "RUF"]
ignore = ["PLR2004", "S101", "COM812", "D205", "D211", "D213", "TRY401", "TRY003", "TRY301"]

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = true
